Abstract observable method to it's own trait.
